requirements.txt
asgiref==3.3.4
Django==3.2.4
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.7.1
flake8==3.9.2
mccabe==0.6.1
pycodestyle==2.7.0
pyflakes==2.3.1
PyJWT==2.1.0
pytz==2021.1
sqlparse==0.4.1
